Handle asset source being a buffer, not a string
I am still trying to track down _why_ this happens, but when doing a production build `source()` returns a `Buffer` instead of a string. It seems to be fine in development, so my intuition is that it's minification related, but I haven't been able to isolate exactly which option. Changing `mode`, disabling the Uglify plugin - nothing seems to alter how this works in a production build.